timeline実装とhelper,decoratorの勉強
