+++ app/assets/stylesheets/_variables.scss
@@ -5,6 +5,8 @@ $gray-font-color-disable: #BDBDBD;
 $black-font-color: #333333; 
 $blue-font-color:#2F80ED;
 
+$usublue-bgc: #b1bdd6;
+
 $btn-yellow-color: #FFE600;
 
 $icon-def-size: 40px;
