+++ app/assets/stylesheets/article.scss
@@ -30,8 +30,9 @@
     
         &-left {
             img {
-                height: $icon-def-size - 10px;
-                width: $icon-def-size -10px;
+
+                height: $icon-def-size - 10;
+                width: $icon-def-size - 10;
                 border-radius: 999px;
             }
         }
