+++ app/assets/stylesheets/tabs.scss
@@ -1,0 +1,33 @@
+@import './variables';
+
+.tabs {
+    display: flex;
+    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.25);
+    // justify-content: space-around;
+    margin: 0 auto 40px;
+    width: 60%;
+
+    &_item {
+        padding: 8px 0px;
+        background-color: white;
+        text-align: center;
+        cursor: pointer;
+        width: 50%;
+
+        a {
+            display: block;
+        }
+
+        &.active {
+            background-color: $usublue-bgc;
+            
+            a {
+                color: white;               
+            }
+        }
+
+        &:hover {
+            opacity: 0.7;
+        }
+    }
+}
