+++ app/controllers/application_controller.rb
@@ -7,6 +7,12 @@ class ApplicationController < ActionController::Base
     before_action :configure_permitted_parameters, if: :devise_controller?
 
 
+    # active decoratorでcurrent_userを使うために
+    def current_user
+        ActiveDecorator::Decorator.instance.decorate(super) if super.present?
+        super
+    end
+
     # private
     def configure_permitted_parameters
         devise_parameter_sanitizer.permit(:sign_up, keys: [ :account])
