+++ app/decorators/article_decorator.rb
@@ -1,0 +1,6 @@
+module  ArticleDecorator
+    def display_created_at
+        # binding.pry
+        I18n.l(self.created_at, format: :default)
+    end
+end
