+++ app/helpers/tabs_helper.rb
@@ -1,0 +1,5 @@
+module TabsHelper
+    def add_active_class(path)
+        'active' if current_page?(path)
+    end
+end
