+++ app/models/article.rb
@@ -20,10 +20,5 @@ class Article < ApplicationRecord
     has_many :comments, dependent: :destroy
     validates :content, presence: true
 
-    
-    def display_created_at
-        # binding.pry
-        I18n.l(self.created_at, format: :default)
-    end
 
 end
