+++ app/views/articles/index.html.haml
@@ -1,62 +1,2 @@
 -# INDEX
-.container
-    .page-insta
-        = image_tag 'insta-mark.png'
-    - @articles.each do |article|
-        .article-item#article-index{ data: { article_id: article.id } }
-            .article-item-wrapper
-                .article-item-left
-                    -# = image_tag 'def-avatar'
-                    = link_to account_path(article.user) do
-                        = image_tag article.user.avatar_image(article.user)
-                .article-item-right
-                    -# %p.article-item-right-name Jmax_21
-                    -# %p.article-item-right-time 3 hours ago
-                    %p.article-item-right-name=article.user.account
-                    -# %p.article-item-right-time=article.updated_at
-                    -# %time(datetime="#{article.created_at}")= time_ago_in_words(article.created_at)
-                    %p.article-item-right-time=article.display_created_at
-            - if article.images.present?
-                .swiper
-                    / Additional required wrapper
-                    .swiper-wrapper
-                        - article.images.each do |image|
-                            / Slides
-                            -# .swiper-slide
-                            -#     = image_tag 'Rectangle.png'
-                            -# .swiper-slide
-                            -#     = image_tag 'Rectangle-1.png'
-                            -# .swiper-slide 
-                            -#     = image_tag 'Rectangle-2.png'
-                            .swiper-slide 
-                                = image_tag image
-                    / If we need pagination
-                    .swiper-pagination
-                    / If we need navigation buttons
-                    .swiper-button-prev
-                    .swiper-button-next
-                    / If we need scrollbar
-                    .swiper-scrollbar
-            .article-item-wrapper.mt-mark
-                .article-item-mark
-                    .article-item-mark-heart.hidden.active-heart
-                        = image_tag 'heart-active.png'
-                    .article-item-mark-heart.hidden.inactive-heart
-                        = image_tag 'heart.png'
-                .article-item-mark
-                    -# = link_to new_article_comment_path(article) do 
-                    = link_to article_comments_path(article) do 
-                        = image_tag 'comment-mark.png'
-                .article-item-mark
-                    -# = image_tag 'link-mark.png'
-                    -# = link_to "https://twitter.com/share?url=#{request.url}" do
-                    = link_to "https://twitter.com/intent/tweet?url=#{request.url}&text=#{article.content}" do
-                        = image_tag 'link-mark.png'
-            .article-item-text
-                -# %p aaaaaaaaaaaaaaaaaaaaaaaa
-                %p= article.content
-    .article-create
-        = link_to new_article_path do
-            %i.fa-solid.fa-circle-plus.fa-3x
-
-=  javascript_pack_tag 'articleIndex'
+= render 'commons/article', articles: @articles
