+++ app/views/commons/_tabs.html.haml
@@ -1,0 +1,5 @@
+.tabs
+    .tabs_item{ class: "#{add_active_class(root_path)}"}
+        = link_to '記事一覧', root_path
+    .tabs_item{ class: "#{add_active_class(timeline_path)}"}
+        = link_to 'タイムライン', timeline_path
