+++ app/views/timelines/show.html.haml
@@ -1,0 +1,2 @@
+-# TimeLine
+= render 'commons/article', articles: @articles
